requirements.txt
crispy-bootstrap5==2024.2
django==5.0.2
django-allauth==0.61.1
django-crispy-forms==2.1
django-environ==0.11.2

# queries to boardgamegeek.com
retry==0.9.2
requests==2.31.0
beautifulsoup4==4.12.3
